Sets Home screen as initial route
Changes the initial route from the Welcome screen to the new Home screen.

Removes the Welcome screen and its associated navigation entry.

Adds a basic Home screen with a placeholder text element.

Persisted state is now only loaded if it exists. This prevents
the app from crashing when there is no persisted state.